error checks

+++ CityFeels/app/src/androidTest/java/com/example/ana/cityfeels/DirectionTest.java
@@ -7,7 +7,7 @@ public class DirectionTest extends AndroidTestCase {
     public void testDireita() {
         Direction direction = new Direction("[ori]", 0);
 
-        String text = direction.applyOrientation(-90);
+        String text = direction.applyOrientation((float)-90);
 
         assertTrue(text.indexOf("À sua direita") == 0);
     }
@@ -15,7 +15,7 @@ public class DirectionTest extends AndroidTestCase {
     public void testEsquerda() {
         Direction direction = new Direction("[ori]", 0);
 
-        String text = direction.applyOrientation(90);
+        String text = direction.applyOrientation((float)90.0);
 
         assertTrue(text.indexOf("À sua esquerda") == 0);
     }
@@ -23,7 +23,7 @@ public class DirectionTest extends AndroidTestCase {
     public void testFrente() {
         Direction direction = new Direction("[ori]", 0);
 
-        String text = direction.applyOrientation(0);
+        String text = direction.applyOrientation((float)0);
 
         assertTrue(text.indexOf("À sua frente") == 0);
     }
@@ -31,7 +31,7 @@ public class DirectionTest extends AndroidTestCase {
     public void testAtras() {
         Direction direction = new Direction("[ori]", 0);
 
-        String text = direction.applyOrientation(180);
+        String text = direction.applyOrientation((float)180);
 
         assertTrue(text.indexOf("Atrás de si") == 0);
     }
